+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 
 ## v1.2.0 (TBA)
 ### General
-* Fix [#176](https://g1cp.org/issues/176): Items with no value no longer show an empty "Value:" in the inventory text.
 ### Story
 
 ## v1.1.0 (TBA)
@@ -15,6 +14,7 @@
 * Fix [#110](https://g1cp.org/issues/110): Mouse input is no longer locked when interacting with objects to allow adjusting the camera viewpoint.
 * Fix [#149](https://g1cp.org/issues/149): The armor "Improved ore Armor" is now correctly labelled as "Improved Ore Armor".
 * Fix [#152](https://g1cp.org/issues/152): The description of the ring "Protection of Fire" is corrected to "Ring of Fire Protection".
+* Fix [#176](https://g1cp.org/issues/176): Items with no value no longer show an empty "Value:" in the inventory text.
 * Fix [#192](https://g1cp.org/issues/192): Mages (NPCs fighting only with spells) no longer auto-equip weapons (e.g. after trading). This requires fix #59 to be active.
 * Fix [#193](https://g1cp.org/issues/193): Objects activated by switches/levers/winches are no longer stuck after loading a game.